etl/requirements.txt
aiohttp==3.9.5
elasticsearch==8.13.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2==2.9.9
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
